python_bot/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
apscheduler>=3.10.4
pyotp>=2.9.0
pydantic>=2.5.0
colorama>=0.4.6
